server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Areeb Khan <areebk@umich.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-generativeai = "^0.5.0"
python-dotenv = "^1.0.1"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
